MALI: Drop use of ksys_close(), rework fd_install usage
GregKH objected to the __close_fd() symbol being used
by the mali driver, via ksys_close(), suggesting I reach
out the the mali team as it was supposedly fixed.

In talking with the mali folks, it is not yet fixed.

The only usage is in an error path in the case that the
copy_to_user fails, where we have an fd that was created
but not passed out to userland.

Greg suggested fput/put_unused_fd() to be used instead
which requires refactoring the logic a bit (as the
kbase_sync_fence_out_create() call would return an
installed fd), so that we can only install the fd
if the copy_to_user succeeds, which matches the pattern
used elsewhere in the kernel.

Careful review would be appreciated!

Bug: 200010226
Test: boot tested on raven
Signed-off-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ia03191a32f8770cf3e9d9b30a13b18e13edb7419